+++ src/Traits/CanDirectCharge.php
@@ -43,6 +43,9 @@ trait CanDirectCharge
      * @param Customer|string|null $customer
      * @param string|null          $orderId
      * @param Metadata|string|null $metadata
+     * @param Basket|null          $basket   The Basket object corresponding to the payment.
+     *                                       The Basket object will be created automatically if it does not exist
+     *                                       yet (i.e. has no id).
      *
      * @return Charge
      *
@@ -55,7 +58,8 @@ trait CanDirectCharge
         $returnUrl,
         $customer = null,
         $orderId = null,
-        $metadata = null
+        $metadata = null,
+        $basket = null
     ): Charge {
         if ($this instanceof HeidelpayParentInterface) {
             return $this->getHeidelpayObject()->charge(
@@ -65,7 +69,8 @@ trait CanDirectCharge
                 $returnUrl,
                 $customer,
                 $orderId,
-                $metadata
+                $metadata,
+                $basket
             );
         }